+++ custom_typings/main.d.ts
@@ -2,3 +2,4 @@
 /// <reference path="./indent.d.ts" />
 /// <reference path="knuth-shuffle.d.ts" />
 /// <reference path="./strip-indent.d.ts" />
+/// <reference path="./whatwg-url.d.ts" />