+++ custom_typings/whatwg-url.d.ts
@@ -1,0 +1,8 @@
+declare module 'whatwg-url' {
+  export class URL {
+    protocol: string;
+    pathname: string;
+
+    constructor(url: string);
+  }
+}