+++ src/test/static/javascript/dynamic-import.js
@@ -1,12 +1,6 @@
-import('./submodule.js').then((submodule) => {
-  console.log(submodule.subThing);
-});
+import('./submodule.js');
+import('test-package');
 
-import('names/not/supported/yet').then((importByName) => {
-  console.log('you know this is not supported yet');
-});
-
+// This is conceivably statically analyzable, but we do not do this yet
 const someString = './module.js';
-import(someString).then((someModule) => {
-  console.log('this is conceivably statically analyzable, but we do not do this yet');
-});
+import(someString);