+++ src/test/static/javascript/module-with-named-import.js
@@ -1,1 +1,1 @@
-import * as foo from 'foo';
+import * as test from 'test-package';